requirements.txt
abstract_singleton==1.0.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
auto_gpt_plugin_template==0.0.3
autoflake==2.1.1
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.3.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
chroma-hnswlib==0.7.2
chromadb==0.4.6
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
confection==0.0.4
coverage==7.2.3
cssselect==1.2.0
cymem==2.0.7
dataclasses-json==0.5.14
distlib==0.3.6
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
exceptiongroup==1.1.1
fastapi==0.99.1
filelock==3.12.0
flake8==6.0.0
flatbuffers==23.5.26
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
gTTS==2.3.1
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httptools==0.6.0
httpx==0.24.0
humanfriendly==10.0
identify==2.5.22
idna==3.4
importlib-resources==6.0.1
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
jq==1.4.1
jsonschema==4.17.3
langchain==0.0.267
langcodes==3.3.0
langsmith==0.0.24
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.20.1
mccabe==0.7.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nodeenv==1.7.0
numexpr==2.8.5
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.15.1
openai==0.27.2
openapi-python-client==0.13.4
openapi-schema-pydantic==1.2.4
orjson==3.8.10
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pathspec==0.11.1
pathy==0.10.1
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.0.0
posthog==3.0.2
pre-commit==3.2.2
preshed==3.0.8
protobuf==4.22.3
pulsar-client==3.2.0
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pyparsing==3.0.9
PyPika==0.48.9
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.9.0
shellingham==1.5.0.post1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.20
srsly==2.4.6
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
thinc==8.1.9
tiktoken==0.3.3
tokenizers==0.13.3
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.22.0
wasabi==0.10.1
watchfiles==0.19.0
webdriver-manager==3.8.6
websocket-client==1.5.1
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.1
